Move LifeCycle classes into wvlet.airframe.lifecycle for the ease of logging
