+++ airframe-di-macros/src/main/scala/wvlet/airframe/AirframeMacros.scala
@@ -493,7 +493,7 @@ private[wvlet] object AirframeMacros {
     q"""{
          val session = ${h.findSession}
          val dep = ${c.prefix}.dep
-         session.lifeCycleManager.addInitHook(wvlet.airframe.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
+         session.lifeCycleManager.addInitHook(wvlet.airframe.lifecycle.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
          dep
         }
       """
@@ -506,7 +506,7 @@ private[wvlet] object AirframeMacros {
     q"""{
          val session = ${h.findSession}
          val dep = ${c.prefix}.dep
-         session.lifeCycleManager.addInjectHook(wvlet.airframe.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
+         session.lifeCycleManager.addInjectHook(wvlet.airframe.lifecycle.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
          dep
         }
       """
@@ -519,7 +519,7 @@ private[wvlet] object AirframeMacros {
     q"""{
          val session = ${h.findSession}
          val dep = ${c.prefix}.dep
-         session.lifeCycleManager.addStartHook(wvlet.airframe.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
+         session.lifeCycleManager.addStartHook(wvlet.airframe.lifecycle.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
          dep
         }
       """
@@ -534,7 +534,7 @@ private[wvlet] object AirframeMacros {
     q"""{
          val session = ${h.findSession}
          val dep = ${c.prefix}.dep
-         session.lifeCycleManager.addStartHook(wvlet.airframe.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
+         session.lifeCycleManager.addStartHook(wvlet.airframe.lifecycle.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
          dep
         }
       """
@@ -547,7 +547,7 @@ private[wvlet] object AirframeMacros {
     q"""{
          val session = ${h.findSession}
          val dep = ${c.prefix}.dep
-         session.lifeCycleManager.addPreShutdownHook(wvlet.airframe.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
+         session.lifeCycleManager.addPreShutdownHook(wvlet.airframe.lifecycle.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
          dep
         }
       """
@@ -560,7 +560,7 @@ private[wvlet] object AirframeMacros {
     q"""{
          val session = ${h.findSession}
          val dep = ${c.prefix}.dep
-         session.lifeCycleManager.addShutdownHook(wvlet.airframe.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
+         session.lifeCycleManager.addShutdownHook(wvlet.airframe.lifecycle.EventHookHolder(${h.surfaceOf(t)}, dep, ${body}))
          dep
         }
       """
