+++ airframe/.js/src/main/scala/wvlet/airframe/lifecycle/JSR250LifeCycleExecutor.scala
@@ -11,10 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe
+package wvlet.airframe.lifecycle
 
-object AddShutdownHook extends LifeCycleEventHandler {
-  override def beforeStart(lifeCycleManager: LifeCycleManager): Unit = {
-    // no-op for Scala.js since there is no shutdown hook
-  }
+/**
+  *
+  */
+object JSR250LifeCycleExecutor extends LifeCycleEventHandler {
+  // Do nothing in Scala.js
 }
