+++ airframe/src/main/scala/wvlet/airframe/lifecycle/LifeCycleEventHandler.scala
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe
+package wvlet.airframe.lifecycle
 
 import wvlet.airframe.surface.Surface
