+++ airframe/src/main/scala/wvlet/airframe/tracing/Tracer.scala
@@ -13,9 +13,10 @@
  */
 package wvlet.airframe.tracing
 
+import wvlet.airframe.lifecycle.Injectee
 import wvlet.airframe.surface.Surface
 import wvlet.airframe.tracing.TraceEvent._
-import wvlet.airframe.{Injectee, Session}
+import wvlet.airframe.Session
 import wvlet.log.LogSupport
 
 /**
