+++ airframe/src/test/scala/wvlet/airframe/AirframeTest.scala
@@ -17,6 +17,7 @@ import java.io.PrintStream
 import java.util.concurrent.atomic.AtomicInteger
 
 import wvlet.airframe.AirframeException.{CYCLIC_DEPENDENCY, MISSING_DEPENDENCY, MISSING_SESSION}
+import wvlet.airframe.lifecycle.{LifeCycleEventHandler, LifeCycleManager}
 import wvlet.log.LogSupport
 import wvlet.airframe.surface.{Primitive, Surface}
 import wvlet.airspec.AirSpec
