+++ airframe/src/test/scala/wvlet/airframe/lifecycle/DesignTimeLifeCycleHookTest.scala
@@ -11,10 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe
+package wvlet.airframe.lifecycle
 
-import java.util.concurrent.atomic.{AtomicBoolean, AtomicInteger}
+import java.util.concurrent.atomic.AtomicInteger
 
+import wvlet.airframe.newSilentDesign
 import wvlet.airspec.AirSpec
 
 /**
