+++ airframe/src/test/scala/wvlet/airframe/lifecycle/LazyStartTest.scala
@@ -11,12 +11,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe
+package wvlet.airframe.lifecycle
 
 import java.util.concurrent.atomic.AtomicBoolean
 
+import wvlet.airframe.{bind, newSilentDesign}
 import wvlet.airspec.AirSpec
 import wvlet.log.LogSupport
+import wvlet.airframe._
 
 object LazyStartTest {
